refactor(v2): modularize Tauri feature adapters
